+++ apps/nextjs-website/src/_contents/pagoPa/guideListsPath.ts
@@ -1,6 +1,0 @@
-import { Path } from '@/lib/types/path';
-
-export const pagoPaGuideListsPath: Path = {
-  name: 'Guide e manuali',
-  path: '/pago-pa/guides',
-};