+++ apps/nextjs-website/src/app/[productSlug]/overview/page.tsx
@@ -7,7 +7,6 @@ import ProductLayout, {
 import { Product } from '@/lib/types/product';
 import { Tutorial } from '@/lib/types/tutorialData';
 import StartInfo from '@/components/organisms/StartInfo/StartInfo';
-import { translations } from '@/_contents/translations';
 import RelatedLinks from '@/components/atoms/RelatedLinks/RelatedLinks';
 import TutorialsOverview from '@/components/organisms/TutorialsOverview/TutorialsOverview';
 import Feature from '@/editorialComponents/Feature/Feature';
@@ -115,7 +114,7 @@ export async function generateMetadata(
     title: product.name,
     description: product.description,
     url: path,
-    image: product.logo.url,
+    image: product.logo?.url,
   });
 }
 
@@ -124,15 +123,14 @@ const OverviewPage = async ({ params }: ProductParams) => {
     hero,
     startInfo,
     feature,
-    product,
     path,
     tutorials,
     postIntegration,
     relatedLinks,
     bannerLinks,
     seo,
+    product,
   } = await getOverview(params.productSlug);
-  const { overview } = translations;
 
   const tutorialsListToShow = tutorials?.list
     ?.filter((tutorial) => tutorial.showInOverview)
@@ -171,23 +169,25 @@ const OverviewPage = async ({ params }: ProductParams) => {
       )}
       {startInfo && (
         <StartInfo
-          title={startInfo.title || overview.startInfo.title}
+          title={startInfo.title}
           cta={startInfo.cta}
           cards={startInfo.cards}
         />
       )}
-      {product.subpaths.tutorials && tutorials && (
+      {product?.hasTutorialListPage && tutorials && (
         <TutorialsOverview
-          title={tutorials.title || overview.tutorial.title}
+          title={tutorials.title}
           subtitle={tutorials.subtitle}
-          ctaLabel={overview.tutorial.ctaLabel}
-          tutorialPath={product.subpaths.tutorials}
+          tutorialPath={{
+            path: `/${product.slug}/tutorials`,
+            name: 'tutorials',
+          }}
           tutorials={[...(tutorialsListToShow || [])]}
         />
       )}
-      {product.subpaths.guides && postIntegration && (
+      {product?.hasGuideListPage && postIntegration && (
         <PostIntegration
-          title={postIntegration.title || overview.postIntegration.title}
+          title={postIntegration.title}
           subtitle={postIntegration.subtitle}
           cta={
             postIntegration.cta && {
@@ -203,10 +203,7 @@ const OverviewPage = async ({ params }: ProductParams) => {
         />
       )}
       {relatedLinks && (
-        <RelatedLinks
-          title={relatedLinks.title || overview.relatedLinks.title}
-          links={relatedLinks.links}
-        />
+        <RelatedLinks title={relatedLinks.title} links={relatedLinks.links} />
       )}
     </ProductLayout>
   );