+++ apps/nextjs-website/src/lib/strapi/fetches/fetchApiDataListPages.ts
@@ -1,6 +1,7 @@
 import * as qs from 'qs';
 import { fetchFromStrapi } from '@/lib/strapi/fetchFromStrapi';
 import { ApiDataListPagesCodec } from '@/lib/strapi/codecs/ApiDataListPagesCodec';
+import { productRelationsPopulate } from './fetchProducts';
 
 const makeStrapiApiDataListPagePopulate = () =>
   qs.stringify({
@@ -15,7 +16,7 @@ const makeStrapiApiDataListPagePopulate = () =>
         },
       },
       product: {
-        populate: ['logo', 'bannerLinks.icon'],
+        ...productRelationsPopulate,
       },
       bannerLinks: {
         populate: ['icon'],